# pkg/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a xssfox api swagger",
        "title": "XSSFox API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:6664",
    "basePath": "/",
    "paths": {
        "/scan": {
            "post": {
                "description": "add xssfox scan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "scan",
                "parameters": [
                    {
                        "description": "json data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.Req"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Res"
                        }
                    }
                }
            }
        },
        "/scan/{scanid}": {
            "get": {
                "description": "get scan info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "scan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "scan id",
                        "name": "scanid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Res"
                        }
                    }
                }
            }
        },
        "/scans": {
            "get": {
                "description": "show scan list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "scan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.Req": {
            "type": "object",
            "properties": {
                "options": {
                    "type": "object",
                    "additionalProperties": true
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "server.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}